Create removeElement.c
This C program is designed to remove all occurrences of a specified integer value from a given array of integers. It prompts the user to enter the number of elements in the array and then to input each element. After obtaining the value to be removed, the program processes the array in-place, shifting the remaining elements forward and counting how many elements are left after the removal. Finally, it displays the count of the remaining elements and prints the modified array, which contains only the elements that are not equal to the specified value.