Create TowerOfHanoi.c
The Tower of Hanoi is a mathematical puzzle or problem involving three rods and a number of disks of different sizes that can slide onto any rod. The puzzle begins with all the disks stacked in decreasing order of size on one rod, with the largest disk at the bottom and the smallest disk at the top.

Objective:
The goal of the puzzle is to move the entire stack of disks from the source rod to the destination rod, following these rules:

Only one disk can be moved at a time.
Each move consists of taking the top disk from one of the stacks and placing it on another rod.
No disk may be placed on top of a smaller disk.
Problem Setup:
There are 3 rods:
Source rod (A) – where all the disks are initially stacked.
Auxiliary rod (B) – used to help transfer disks.
Destination rod (C) – where the disks need to be moved.
Number of disks: n.

Steps to Solve the Puzzle:
Move the top n-1 disks from the source rod (A) to the auxiliary rod (B), using the destination rod (C) as a temporary holding area.
Move the nth (largest) disk directly from the source rod (A) to the destination rod (C).
Finally, move the n-1 disks from the auxiliary rod (B) to the destination rod (C), using the source rod (A) as an auxiliary.
Recursive Solution:
The problem can be solved recursively. Each step breaks down the problem into a smaller instance of the same problem (moving n-1 disks) until we reach the base case where there's only one disk left to move.

The steps for solving Tower of Hanoi using recursion are:
1) Move n-1 disks from source rod A to auxiliary rod B.
2) Move the nth disk from source rod A to destination rod C.
3) Move n-1 disks from auxiliary rod B to destination rod C.