Create 015_ElementInsertionArrayInBetween.c
Program: Array Element Insertion in C
This program allows the user to insert a new element into an array at a specified index. The user is prompted to enter the current elements of the array, the index where they want to insert the new element, and the value to be inserted. After the insertion, the program prints the updated array with the new element included.