Create 016.ElementDeletionArray.c
Program: Array Element Deletion in C
This program performs element deletion from an array based on a user-specified index. The user is prompted to input the current elements of the array and the index of the element to be deleted. The program then shifts the elements of the array to the left, effectively removing the element at the specified index. Finally, the updated array is displayed with the element removed.