Create 017.LinearSearch.c
Program: Linear Search in C
This program implements a linear search algorithm to find a specific element in an array. The user inputs the number of elements and the elements themselves. Then, the user specifies the element they want to search for. The program scans the array and prints the index of the element if found, otherwise it displays a message that the element was not found. This simple search algorithm checks each element sequentially, making it easy to understand and implement.